added a class file for importing object classes and created basics of publish-item
