+++ src/app/short-detail/publish-item/publish-item.component.html
@@ -1,3 +1,5 @@
 <p>
-  publish-item works!
+  {{publishItem.verNo}}
+  {{publishItem.publisher}}
+  {{publishItem.pbCmt}}
 </p>
