+++ src/app/short-detail/publish-item/publish-item.component.ts
@@ -1,11 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
+import {PublishHistoryList} from '../../object-classes/service-classes';
 
 @Component({
   selector: 'app-publish-item',
   templateUrl: './publish-item.component.html',
-  styleUrls: ['./publish-item.component.css']
+  styleUrls: ['publish-item.component.scss']
 })
 export class PublishItemComponent implements OnInit {
+
+  @Input() publishItem: PublishHistoryList;
 
   constructor() { }
