+++ src/app/short-detail/short-detail.component.scss
@@ -1,0 +1,12 @@
+.publishItemContainer{
+  display: flex;
+  flex-flow: row wrap;
+}
+
+.publishItem{
+  margin-right: 10px;
+  border: dotted;
+  border-width: 2px;
+  border-color: #548264;
+}
+
