+++ src/app/short-detail/short-detail.component.ts
@@ -3,23 +3,13 @@ import {ActivatedRoute, Params} from '@angular/router';
 import {PublishHistoryService} from '../publish-history.service';
 import {Observable} from "rxjs";
 import 'rxjs/add/operator/switchMap';
-
-class PublishHistoryList {
-  constructor(public verNo: string,
-              public taskType: number,
-              public publisher: string,
-              public pbDate: string,
-              public pbCmt: string) {
-  }
-}
-
-
+import {PublishHistoryList} from '../object-classes/service-classes'
 
 
 @Component({
   selector: 'app-short-detail',
   templateUrl: './short-detail.component.html',
-  styleUrls: ['./short-detail.component.css'],
+  styleUrls: ['short-detail.component.scss'],
   providers: [PublishHistoryService]
 })
 export class ShortDetailComponent implements OnInit {
@@ -33,7 +23,6 @@ export class ShortDetailComponent implements OnInit {
   ngOnInit() {
     this.pubHistoryList = this.route.params.switchMap((params: Params) =>{
       this.id = params['seq'];
-      console.log('hello');
       return this.pubHistoryService.getPubList(this.id);
     });
   }
