+++ src/app/shot.service.ts
@@ -1,27 +1,9 @@
 import {Injectable} from '@angular/core';
 import {Http, Response, URLSearchParams} from '@angular/http';
-
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
+import {Short} from './object-classes/service-classes';
 
-class Sequence {
-  constructor(public seq: number,
-              public id: string,
-              public name: string) {
-  }
-}
-
-class nestedShort {
-  constructor(public seq: number,
-              public shotID: string,
-              public shotName: string) {
-  }
-}
-class Short {
-  constructor(public sequenceSequence: number,
-              public shorts: nestedShort) {
-  }
-}
 
 @Injectable()
 export class ShortService {
